pyproject.toml
[tool.poetry]
name = "poetry-commerce"
version = "0.1.0"
description = "skip"
authors = ["Pavel Bobrov"]
license = "skip"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pydantic = "^2.6.1"
httpx = "^0.27.0"
fastapi-cache2 = "^0.2.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.8.0"
pytest = "^8.0.1"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
